@@ -1,8 +1,8 @@
 module webhook-go
 
-go 1.13
+go 1.16
 
 require (
 	github.com/axgle/mahonia v0.0.0-20180208002826-3358181d7394
-	golang.org/x/crypto v0.0.0-20191227163750-53104e6ec876
+	golang.org/x/crypto v0.0.0-20210322153248-0c34fe9e7dc2
 )
